+++ tortoise/backends/base/executor.py
@@ -114,10 +114,15 @@ class BaseExecutor:
                 for model, index, *__, full_path in self.select_related_idx[1:]:
                     (*path, attr) = full_path
                     related_items = row_items[current_idx : current_idx + index]
-                    if not any((v for _, v in related_items)):
-                        obj = None
-                    else:
+                    if any(v for _, v in related_items):
                         obj = model._init_from_db(**{k.split(".")[1]: v for k, v in related_items})
+                    elif index == 0:
+                        # 0 signals that an empty "filler" object should be created in the case
+                        # where a field of related model is selected but model itself isn't,
+                        # e.g. .only("relatedmodel__field")
+                        obj = model._init_from_db()
+                    else:
+                        obj = None
                     target = instances.get(tuple(path))
                     if target is not None:
                         setattr(target, f"_{attr}", obj)