+++ tortoise/query_utils.py
@@ -2,11 +2,12 @@ from __future__ import annotations
 
 from copy import copy
 from typing import TYPE_CHECKING, cast
+from collections.abc import Sequence
 
 from pypika_tortoise import Table
 from pypika_tortoise.terms import Criterion, Term
 
-from tortoise.exceptions import ConfigurationError, OperationalError
+from tortoise.exceptions import FieldError, OperationalError
 from tortoise.fields.base import Field
 from tortoise.fields.relational import (
     BackwardFKRelation,
@@ -78,6 +79,24 @@ def get_joins_for_related_field(
     return required_joins
 
 
+def expand_lookup_expression(root_model: type[Model], lookup_expression: str) -> Sequence[Field]:
+    field_names = lookup_expression.split("__")
+    fields: list[Field | RelationalField] = []
+    model = root_model
+    for field_name in field_names[:-1]:
+        if field_name not in model._meta.fetch_fields:
+            raise FieldError(f"{lookup_expression} not resolvable")
+        field = cast(RelationalField, model._meta.fields_map[field_name])
+        fields.append(field)
+        model = field.related_model
+    # the last field is not necessarily a RelationalField, so threting it differently
+    try:
+        fields.append(model._meta.fields_map[field_names[-1]])
+    except KeyError:
+        raise FieldError(f"{lookup_expression} not resolvable")
+    return fields
+
+
 def resolve_nested_field(
     model: type[Model], table: Table, field: str
 ) -> tuple[Term, list[TableCriterionTuple], Field | None]:
@@ -86,52 +105,51 @@ def resolve_nested_field(
     returns the pypika term, required joins and the Field that can be used for
     converting the value.
     """
-    field_object = None
     joins = []
-    fields = field.split("__")
+    fields = expand_lookup_expression(model, field)
 
     for iter_field in fields[:-1]:
-        if iter_field not in model._meta.fetch_fields:
-            raise ConfigurationError(f"{field} not resolvable")
-
-        related_field = cast(RelationalField, model._meta.fields_map[iter_field])
-        joins.extend(get_joins_for_related_field(table, related_field, iter_field))
+        related_field = cast(RelationalField, iter_field)
+        joins.extend(get_joins_for_related_field(table, related_field, iter_field.model_field_name))
 
         model = related_field.related_model
         related_table: Table = related_field.related_model._meta.basetable
         if isinstance(related_field, ForeignKeyFieldInstance):
             # Only FK's can be to same table, so we only auto-alias FK join tables
-            related_table = related_table.as_(f"{table.get_table_name()}__{iter_field}")
+            related_table = related_table.as_(
+                f"{table.get_table_name()}__{iter_field.model_field_name}"
+            )
         table = related_table
 
     last_field = fields[-1]
-    if last_field in model._meta.fetch_fields:
-        related_field = cast(RelationalField, model._meta.fields_map[last_field])
+    if last_field.model_field_name in model._meta.fetch_fields:
+        related_field = cast(RelationalField, last_field)
         related_field_meta = related_field.related_model._meta
 
-        joins.extend(get_joins_for_related_field(table, related_field, last_field))
+        joins.extend(
+            get_joins_for_related_field(table, related_field, related_field.model_field_name)
+        )
         related_table = related_field_meta.basetable
 
         if isinstance(related_field, BackwardFKRelation):
             if table == related_table:
-                related_table = related_table.as_(f"{table.get_table_name()}__{last_field}")
+                related_table = related_table.as_(
+                    f"{table.get_table_name()}__{related_field.model_field_name}"
+                )
 
         term = related_table[related_field_meta.db_pk_column]
     else:
-        field_object = model._meta.fields_map[last_field]
-        if field_object.source_field:
-            term = table[field_object.source_field]
+        if last_field.source_field:
+            term = table[last_field.source_field]
         else:
-            term = table[last_field]
+            term = table[last_field.model_field_name]
 
-        if field_object:  # pragma: nobranch
-            func = field_object.get_for_dialect(
-                model._meta.db.capabilities.dialect, "function_cast"
-            )
+        if last_field:  # pragma: nobranch
+            func = last_field.get_for_dialect(model._meta.db.capabilities.dialect, "function_cast")
             if func:
-                term = func(field_object, term)
+                term = func(last_field, term)
 
-    return term, joins, field_object
+    return term, joins, last_field
 
 
 class EmptyCriterion(Criterion):